+++ src/app/admin/posts/[id]/page.tsx
@@ -2,21 +2,31 @@
 import { Category } from "@/types/Category";
 import { useParams, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
+
+
 import { Post } from "@/types/Post";
 import { PostForm } from "../_components/PostForm";
 import { UpdatePostRequestBody } from "@/app/api/admin/posts/[id]/route";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
 
 export default function Page(){
   const [title,setTitle] = useState<string>("")
   const [content,setContent] = useState<string>("")
-  const [thumbnailUrl,setThumbnailUrl] = useState<string>("https://placehold.jp/800x400.png")
+  const [thumbnailUrl,setThumbnailUrl] = useState<string>("")
   const [categories, setCategories] = useState<Category[]>([])
   const router = useRouter();
   const {id} = useParams();
+  const {token} = useSupabaseSession()
+
 
   useEffect(() => {
+    if(!token) return;
     const fetcher = async () => {
-      const res = await fetch(`/api/admin/posts/${id}`)
+      const res = await fetch(`/api/admin/posts/${id}`,{
+        headers:{
+          Authorization: token,
+        }
+      })
       const {post}:{post:Post} = await res.json();
       setTitle(post.title)
       setContent(post.content)
@@ -24,13 +34,20 @@ export default function Page(){
       setCategories(post.postCategories.map((pc) => pc.category))
     }
     fetcher();
-  },[id])
+  },[id,token])
+
+
 
   const handleDelete = async () => {
+    if(!token) return;
+
     if(!confirm("本当に削除しますか？")) return;
 
     const res = await fetch(`/api/admin/posts/${id}`,{
-      method:"DELETE"
+      method:"DELETE",
+      headers:{
+        Authorization: token,
+      }
     })
     if(res.status === 200){
       alert('記事を削除しました。')
@@ -43,6 +60,7 @@ export default function Page(){
 
   const handleSubmit = async (e:React.FormEvent) => {
     e.preventDefault();
+    if(!token) return;
 
     if(title.trim() === ""){
       return;
@@ -53,6 +71,7 @@ export default function Page(){
       'method':'PUT',
       headers:{
         'Content-Type': 'application/json',
+        Authorization: token,
       },
       body:JSON.stringify(body),
     })
